full asset list and delegat/undelegate actions (#25)
* Delegate and undelegate UI (#20)

* add delegate ui

* add undelegate ui

* Undelegate claim effect hash and UI (#21)

* add parser

* add effect hash

* add ui

* add flag to remove con in rust

* fix warnings

* update zxlib and ensure the code is compilable

* Feat/all assets (#23)

* Update error definitions

* Add missing assets and adjust max value lengths

* Normalice error definitions

* Remove duplicated asses

* Add balance struct  (#22)

* add balance struct

* fix warning in rust

* Delegator vote (#24)

* fix note for spend plan

* add parser

* add rust interface

* effect hash computation

* add ui

---------

+++ app/rust/src/parser/nullifier.rs
@@ -30,7 +30,7 @@ impl Nullifier {
     pub fn derive(nk: &NullifierKey, pos: u64, state_commitment: &Fq) -> Nullifier {
         Nullifier(hash_3(
             &Self::nullifier_domain_sep(),
-            (nk.0, *state_commitment, (u64::from(pos)).into()),
+            (nk.0, *state_commitment, pos.into()),
         ))
     }
